+++ CHANGELOG.md
@@ -2,8 +2,18 @@
 
 All notable changes to the "nicksays" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
 ## [Unreleased]
 
 - Initial release
+
+## [0.0.3] - 2023-01-30
+
+### Added
+
+- `nicksays.number_cursors_from_one`
+- `nicksays.right_align_cursors_using_spaces`
+
+### Changed
+
+- renamed `nicksays.left_align_cursors_using_spaces`
+- renamed `Number Cursors from Zero`